Automatically add routeIds inside DataStaticRouter
