+++ examples/ssr-data-router/src/entry.server.tsx
@@ -6,7 +6,7 @@ import { DataStaticRouter } from "react-router-dom/server";
 import { routes } from "./App";
 
 export async function render(request: express.Request) {
-  let { dataRoutes, query } = unstable_createStaticHandler(routes);
+  let { query } = unstable_createStaticHandler(routes);
   let remixRequest = createFetchRequest(request);
   let context = await query(remixRequest);
 
@@ -16,11 +16,7 @@ export async function render(request: express.Request) {
 
   return ReactDOMServer.renderToString(
     <React.StrictMode>
-      <DataStaticRouter
-        dataRoutes={dataRoutes}
-        context={context}
-        nonce="the-nonce"
-      />
+      <DataStaticRouter routes={routes} context={context} nonce="the-nonce" />
     </React.StrictMode>
   );
 }
